requirements/dev.txt
appdirs==1.4.4
asgiref==3.4.1
Babel==2.9.1
black==21.6b0
click==8.0.1
Django==3.2.5
django-environ==0.4.5
django-phonenumber-field==4.0.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.4.0
mypy-extensions==0.4.3
pathspec==0.8.1
phonenumberslite==8.12.6
Pillow==8.3.1
PyJWT==1.7.1
pytz==2021.1
regex==2021.7.6
sqlparse==0.4.1
toml==0.10.2
